refactor(api): move branding + static-branding mount into the daemon
Commit 6 of the dashboard fold-in. Lifts the last three /api/*
endpoints out of dashboard/api/main.py:

- GET  /api/projects/{id}/branding
- POST /api/projects/{id}/branding
- POST /api/projects/{id}/branding/logo (Pillow validation, 256 KB / 512x512 OQ4 locks)

into server/api/dashboard/branding.py. The same module owns
mount_static_branding(app), which the daemon's _configure_routers
calls after every router lands so uploaded logos resolve at
http://<daemon>/static/branding/<id>.<ext>. The mount lives at the
FastAPI app level (StaticFiles can't be added through APIRouter).

_BRANDING_DIR resolves to src/dashboard/api/static/branding/ via a
walk up the package tree, matching the legacy path so existing
uploaded files keep working. The OQ4 constants (256 KB cap, 512 px
cap, PNG/SVG-only) carry over verbatim. The Pillow ImportError ->
503 guard is preserved, so the daemon does not hard-depend on
Pillow when voicegateway[dashboard] is not installed.

After this commit, dashboard/api/main.py contains only the legacy
SPA-serving block. Commit 7 (cut-over) moves that into
server/static.py and drops the file.

test_branding_endpoint.py replaces its _FakeGateway shim with a
real Gateway built from temp_config; the daemon's build_app needs
the layered DI stack the fake didn't model.

ruff check src + uv run mypy + uv run pytest -q all green:
1666 passed, 4 skipped.
